Started implementing arrangment files
- Reading files in javascript is a pain due to security restrictions and
browser specific implementations. Hence, all the "arrangements" are
created at runtime and stored in memory. The 'arrangements' directory
still contains an initial json attempt which is not used.
- Currently, the system reads from the 'arrangements' object to load the
simulation, albeit in a very hard-coded way.
- Added a dropdown list of arrangements to the control panel. The
default is the "Earth-Moon System" but a examplary "Sun-Earth System"
was added too (Not implemented yet).
- Added two button icons, not used yet.
- Minor style changes.